+++ src/FonteMassCHK.cpp
@@ -145,7 +145,7 @@ void fontemaschk::VMas(){
 		double masChk=chk.vazmassSachd(ypres,pfonte, tfonte,titfonte,betfonte,titfonte, fluidoPfonte,fluidocol);
 		maxSup=chk.vazmaxSachd(pfonte, tfonte, titfonte,betfonte,titfonte, fluidoPfonte,fluidocol);
 		int fluxcri=1;
-		if(titfonte<=0.01 || fabsl(ypres)>fabsl(chk.razpres)){
+		if(titfonte<=0.01 || fabs(ypres)>fabs(chk.razpres)){
 	         	 fluxcri=0;
 	         	 maxSup=masChk;
 		}
