+++ src/PorosoRad.cpp
@@ -653,7 +653,7 @@ double PorosRad::transperm(double mastot){
 	  double keq=celula[0].kabsol1;
 	  double visc=flup.ViscOleo(1., tRes)/1000.;
 	  double ip=98066.52*(keq/(visc))/
-			         (logl(celula[ncel-1].rm/celula[0].rm)+0.);
+			         (log(celula[ncel-1].rm/celula[0].rm)+0.);
 	  mastot=ip*(celula[0].presRes-celula[0].Pint)*1000.;
 	}
 	double val=marchaperm(mastot);
@@ -805,7 +805,7 @@ void PorosRad::renovaPres(int i, double mTot){
 			if(celula[i].sW>celula[i].sL-1e-15)celula[i].sW=celula[i].sL;
 			else if(celula[i].sW<celula[i].satConata-1e-15)celula[i].sW=celula[i].satConata;
 
-			erro=(fabsl(celula[i].sL-sLiter)+fabsl(celula[i].sW-sWiter));
+			erro=(fabs(celula[i].sL-sLiter)+fabs(celula[i].sW-sWiter));
 
 			kO=celula[i+1].kmed0=celula[i+1].fkO(0.5*(celula[i+1].sW+celula[i].sW), 1.-0.5*(celula[i+1].sL+celula[i].sL));
 			kG=celula[i+1].kmedG0=celula[i+1].interpolaTabela(celula[i+1].kRelOGCel.npont, 1.-0.5*(celula[i+1].sL+celula[i].sL),
@@ -825,7 +825,7 @@ void PorosRad::renovaPres(int i, double mTot){
 
 
 double PorosRad::SIGN(double a, double b) {
-  return (b >= 0 ? 1.0 : -1.0) * fabsl(a);
+  return (b >= 0 ? 1.0 : -1.0) * fabs(a);
 }
 
 
@@ -836,7 +836,7 @@ double  PorosRad::zriddr(double x1,double x2){
 	double xmin;
 	double fl=marchaperm(x1);
 	double fh=marchaperm(x2);
-	if(fabsl(fh)<fabsl(fl)){
+	if(fabs(fh)<fabs(fl)){
 		fmin=fh;
 		xmin=x2;
 	}
@@ -851,27 +851,27 @@ double  PorosRad::zriddr(double x1,double x2){
 		for(int j=0;j<maxit;j++){
 			double xm=0.5*(xl+xh);
 			double fm=marchaperm(xm);
-			if(fabsl(fm)<fabsl(fmin)){
+			if(fabs(fm)<fabs(fmin)){
 				fmin=fm;
 				xmin=xm;
 			}
-			double s=sqrtl(fm*fm-fl*fh);
+			double s=sqrt(fm*fm-fl*fh);
 			if(s==0.0){
 				fmin=marchaperm(xmin);
 				return xmin;
 			}
 			double xnew=xm+(xm-xl)*((fl>=fh ? 1.0 : -1.0)*fm/s);
-			if(fabsl(xnew-ans)<=xacc ){
+			if(fabs(xnew-ans)<=xacc ){
 				fmin=marchaperm(xmin);
 				return xmin;
 			}
 			ans=xnew;
 			double fnew=marchaperm(ans);
-			if(fabsl(fnew)<fabsl(fmin)){
+			if(fabs(fnew)<fabs(fmin)){
 				fmin=fnew;
 				xmin=ans;
 			}
-			if(fabsl(fnew)<=xacc ){
+			if(fabs(fnew)<=xacc ){
 				fmin=marchaperm(xmin);
 				return xmin;
 			}
@@ -890,7 +890,7 @@ double  PorosRad::zriddr(double x1,double x2){
 				fl=fnew;
 			}
 			else return -1.e10;
-			if(fabsl(xh-xl)<=xacc ){
+			if(fabs(xh-xl)<=xacc ){
 				fmin=marchaperm(xmin);
 				return xmin;
 			}
@@ -899,8 +899,8 @@ double  PorosRad::zriddr(double x1,double x2){
 		 return 1.e10;
 	}
 	else {
-		if(fabsl(fl)<=xacc) return x1;
-		if(fabsl(fh)<=xacc) return x2;
+		if(fabs(fl)<=xacc) return x1;
+		if(fabs(fh)<=xacc) return x2;
 		return -1e10;
 	}
 }
@@ -1128,9 +1128,9 @@ double PorosRad::transtrans(double espessura){
 	 matglob.GaussElimPP(vetliv);
 
 
-	// erro+=fabsl(vetliv[0]-Pcamada[0][0]);
+	// erro+=fabs(vetliv[0]-Pcamada[0][0]);
 	 idisc=0;
-	 erro+=fabsl(vetliv[0]-celula[idisc].Pcamada);
+	 erro+=fabs(vetliv[0]-celula[idisc].Pcamada);
 	 //Pcamada[0][0]=vetliv[0];
 	 celula[idisc].Pcamada=vetliv[0];
 	 celula[idisc].PcamadaL=vetliv[0];
@@ -1148,8 +1148,8 @@ double PorosRad::transtrans(double espessura){
 	 //for(int i=0;i<geom.ncamadas;i++){
 		 //for(int j=1;j<=ncamada[i];j++){
 			 //idisc++;
-			 //erro+=fabsl(vetliv[konta]-Pcamada[i][j]);
-			 erro+=fabsl(vetliv[konta]-celula[i].Pcamada);
+			 //erro+=fabs(vetliv[konta]-Pcamada[i][j]);
+			 erro+=fabs(vetliv[konta]-celula[i].Pcamada);
 			 //Pcamada[i][j]=vetliv[konta];
 			 celula[i].Pcamada=vetliv[konta];
 			 celula[i-1].PcamadaR=vetliv[konta];
@@ -1188,7 +1188,7 @@ double PorosRad::transtrans(double espessura){
  double qaSTD=vetliv[3]*celula[0].flup.MasEspAgua(Pint, tRes)/(flup.Denag*1000.);
  double qoSTD;
  qoSTD=vetliv[1]*(1-alf)/flup.BOFunc(Pint, tRes);
- if((fabsl(qaSTD)+fabsl(qoSTD))>1e-15) celula[0].BSW=BSW=fabsl(qaSTD)/(fabsl(qaSTD)+fabsl(qoSTD));
+ if((fabs(qaSTD)+fabs(qoSTD))>1e-15) celula[0].BSW=BSW=fabs(qaSTD)/(fabs(qaSTD)+fabs(qoSTD));
 
  double pAux=0.;
  double qaSTD2;
@@ -1201,8 +1201,8 @@ double PorosRad::transtrans(double espessura){
 	  double alf=0*tit+1*(tit/rhogP)/((tit/rhogP)+((1.-tit)/rhoP));
 	  qaSTD2=celula[i].QwcamadaR*celula[i].flup.MasEspAgua(pAux, tRes)/(flup.Denag*1000.);
 	  qoSTD2=celula[i].QocamadaR*(1-alf)/celula[i].flup.BOFunc(pAux, tRes);
-	  if((fabsl(qaSTD2)+fabsl(qoSTD2))>1e-15){
-		  celula[i].BSW=fabsl(qaSTD2)/(fabsl(qaSTD2)+fabsl(qoSTD2));
+	  if((fabs(qaSTD2)+fabs(qoSTD2))>1e-15){
+		  celula[i].BSW=fabs(qaSTD2)/(fabs(qaSTD2)+fabs(qoSTD2));
 	  }
  }
  BSW=celula[2].BSW;
@@ -1210,7 +1210,7 @@ double PorosRad::transtrans(double espessura){
 
   /*double qaSTD=vetliv[3]*celula[0].flup.MasEspAgua(Pint, tRes)/(flup.Denag*1000.);
   double qoSTD=vetliv[1]*(1-celula[0].alf)/flup.BOFunc(Pint, tRes);
-  if(fabsl(qaSTD+qoSTD)>1e-15) flup.BSW=qaSTD/(qaSTD+qoSTD);*/
+  if(fabs(qaSTD+qoSTD)>1e-15) flup.BSW=qaSTD/(qaSTD+qoSTD);*/
   /*fluxIni=2.*M_PI*vetliv[1]*(1-celula[0].alf)*celula[0].flup.MasEspoleo(Pint, tRes)*espessura;
   fluxIniG=2.*M_PI*vetliv[1]*celula[0].alf*celula[0].flup.MasEspGas(Pint, tRes)*espessura;
   fluxIniA=2.*M_PI*vetliv[3]*celula[0].flup.MasEspAgua(Pint, tRes)*espessura;*/
